Explicitly including <string> OpenThingsFramework.cpp
In DEBUG mode, there is a reference to std::string.  Depending on what version of the espressif8266 version being compiled against, this may not be available.  In particular, std::string is included in the 2.x build toolchain, but not in the 3.x or 4.x, so it causes a compile error unless it's explicitly included.